+++ api/specs/web-server/_licensed_items.py
@@ -14,8 +14,10 @@ from models_library.api_schemas_webserver.licensed_items import LicensedItemGet
 from models_library.generics import Envelope
 from models_library.rest_error import EnvelopedError
 from simcore_service_webserver._meta import API_VTAG
-from simcore_service_webserver.licenses._exceptions_handlers import _TO_HTTP_ERROR_MAP
-from simcore_service_webserver.licenses._models import (
+from simcore_service_webserver.licenses._common.exceptions_handlers import (
+    _TO_HTTP_ERROR_MAP,
+)
+from simcore_service_webserver.licenses._common.models import (
     LicensedItemsBodyParams,
     LicensedItemsListQueryParams,
     LicensedItemsPathParams,
