+++ services/web/server/src/simcore_service_webserver/licenses/_common/models.py
@@ -18,7 +18,7 @@ from models_library.wallets import WalletID
 from pydantic import BaseModel, ConfigDict, Field
 from servicelib.request_keys import RQT_USERID_KEY
 
-from .._constants import RQ_PRODUCT_KEY
+from ..._constants import RQ_PRODUCT_KEY
 
 _logger = logging.getLogger(__name__)
