+++ services/web/server/src/simcore_service_webserver/licenses/_licensed_items_rest.py
@@ -22,8 +22,8 @@ from ..login.decorators import login_required
 from ..security.decorators import permission_required
 from ..utils_aiohttp import envelope_json_response
 from . import _licensed_items_service
-from ._exceptions_handlers import handle_plugin_requests_exceptions
-from ._models import (
+from ._common.exceptions_handlers import handle_plugin_requests_exceptions
+from ._common.models import (
     LicensedItemsBodyParams,
     LicensedItemsListQueryParams,
     LicensedItemsPathParams,
